requirements.txt
pandas
numpy
scikit-learn
nltk
mlxtend
matplotlib
seaborn
pyyaml
papermill
jupyter